micro_services/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
requests
